# docs/api/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Standardized API error message container.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "error_code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HazardEventResponse": {
        "description": "Public hazard event schema.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "event_date": {
            "format": "date-time",
            "title": "Event Date",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "region_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Label"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "id",
          "event_type",
          "region_label",
          "event_date",
          "details",
          "source",
          "external_id",
          "created_at"
        ],
        "title": "HazardEventResponse",
        "type": "object"
      },
      "InvestigationCreateRequest": {
        "description": "Payload to create a new investigation query.",
        "properties": {
          "consent_public_research": {
            "default": false,
            "description": "Explicit opt-in consent to include anonymized investigation query text in public research dataset (ADR-504).",
            "title": "Consent Public Research",
            "type": "boolean"
          },
          "query": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "InvestigationCreateRequest",
        "type": "object"
      },
      "InvestigationCreateResponse": {
        "description": "Payload returned upon successfully queuing an investigation.",
        "properties": {
          "investigation_id": {
            "format": "uuid",
            "title": "Investigation Id",
            "type": "string"
          },
          "poll_url": {
            "title": "Poll Url",
            "type": "string"
          },
          "status": {
            "const": "accepted",
            "title": "Status",
            "type": "string"
          },
          "stream_url": {
            "title": "Stream Url",
            "type": "string"
          }
        },
        "required": [
          "investigation_id",
          "status",
          "stream_url",
          "poll_url"
        ],
        "title": "InvestigationCreateResponse",
        "type": "object"
      },
...
      "PatternFindingResponse": {
        "description": "Public research longitudinal pattern finding schema.",
        "properties": {
          "algorithm_version": {
            "title": "Algorithm Version",
            "type": "string"
          },
          "baseline_rate": {
            "title": "Baseline Rate",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lift": {
            "title": "Lift",
            "type": "number"
          },
          "mined_at": {
            "format": "date-time",
            "title": "Mined At",
            "type": "string"
          },
          "observed_rate": {
            "title": "Observed Rate",
            "type": "number"
          },
          "pattern_hash": {
            "title": "Pattern Hash",
            "type": "string"
          },
          "region_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Label"
          },
          "source_event_type": {
            "title": "Source Event Type",
            "type": "string"
          },
          "statistical_confidence": {
            "title": "Statistical Confidence",
            "type": "number"
          },
          "support_count": {
            "title": "Support Count",
            "type": "integer"
          },
          "supporting_event_ids": {
            "items": {
              "type": "string"
            },
            "title": "Supporting Event Ids",
            "type": "array"
          },
          "target_event_type": {
            "title": "Target Event Type",
            "type": "string"
          },
          "time_window_days": {
            "title": "Time Window Days",
            "type": "integer"
          },
          "total_source_events": {
            "title": "Total Source Events",
            "type": "integer"
          },
          "total_target_events": {
            "title": "Total Target Events",
            "type": "integer"
          },
          "uncertainty": {
            "additionalProperties": true,
            "title": "Uncertainty",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "pattern_hash",
          "algorithm_version",
          "version",
          "source_event_type",
          "target_event_type",
          "region_label",
          "time_window_days",
          "support_count",
          "total_source_events",
          "total_target_events",
          "observed_rate",
          "baseline_rate",
          "lift",
          "statistical_confidence",
          "uncertainty",
          "supporting_event_ids",
          "description",
          "mined_at",
          "created_at"
        ],
        "title": "PatternFindingResponse",
        "type": "object"
      },
...
      "ResearchInvestigationResponse": {
        "description": "Anonymized investigation schema returned to public research consumers.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "complexity_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complexity Tier"
          },
          "confidence_summary": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Summary"
          },
          "consent_public_research": {
            "title": "Consent Public Research",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domains_involved": {
            "items": {
              "type": "string"
            },
            "title": "Domains Involved",
            "type": "array"
          },
          "execution_trace": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Trace"
          },
          "investigation_id": {
            "format": "uuid",
            "title": "Investigation Id",
            "type": "string"
          },
          "query_category": {
            "title": "Query Category",
            "type": "string"
          },
          "query_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Text"
          }
        },
        "required": [
          "investigation_id",
          "query_category",
          "domains_involved",
          "complexity_tier",
          "confidence_summary",
          "consent_public_research",
          "created_at"
        ],
        "title": "ResearchInvestigationResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "An Agentic Planetary Risk Intelligence Platform. Phase 1 — foundational API skeleton.",
    "title": "GaiaOS",
    "version": "0.5.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/investigations": {
      "post": {
        "description": "Create an investigation and invoke the compiled LangGraph asynchronously.",
        "operationId": "create_investigation_api_v1_investigations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Redis Checkpointer unavailable"
          }
        },
        "summary": "Create Investigation",
        "tags": [
          "v1",
          "Investigations"
        ]
      }
    },
...
    "/api/v1/research/hazard-events": {
      "get": {
        "description": "Returns historical hazard events from public data sources (USGS, NOAA, FIRMS).",
        "operationId": "list_public_hazard_events_api_v1_research_hazard_events_get",
        "parameters": [
          {
            "description": "Filter by event type (e.g., earthquake)",
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type (e.g., earthquake)",
              "title": "Event Type"
            }
          },
          {
            "description": "Filter by source (e.g., USGS)",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source (e.g., USGS)",
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HazardEventResponse"
                  },
                  "title": "Response List Public Hazard Events Api V1 Research Hazard Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Public research API disabled"
          }
        },
        "summary": "List public hazard events",
        "tags": [
          "v1",
          "Public Research"
        ]
      }
    },
    "/api/v1/research/investigations": {
      "get": {
        "description": "Returns aggregated, anonymized investigation findings. User identity (user_id) is strictly omitted (ADR-504). Raw query text is included ONLY if the submitter provided explicit consent.",
        "operationId": "list_public_research_investigations_api_v1_research_investigations_get",
        "parameters": [
          {
            "description": "Filter by domain (e.g., seismic)",
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by domain (e.g., seismic)",
              "title": "Domain"
            }
          },
          {
            "description": "Filter investigations since timestamp",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter investigations since timestamp",
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResearchInvestigationResponse"
                  },
                  "title": "Response List Public Research Investigations Api V1 Research Investigations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Public research API disabled"
          }
        },
        "summary": "List anonymized public research findings",
        "tags": [
          "v1",
          "Public Research"
        ]
      }
    },
    "/api/v1/research/patterns": {
      "get": {
        "description": "Returns statistically-notable recurring co-occurrence patterns discovered across historical hazard events. Findings are computed periodically by background jobs and versioned to track historical confidence trends.",
        "operationId": "list_public_research_patterns_api_v1_research_patterns_get",
        "parameters": [
          {
            "description": "Filter by source or target event type (e.g. earthquake)",
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source or target event type (e.g. earthquake)",
              "title": "Event Type"
            }
          },
          {
            "description": "Filter by geographic region label",
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by geographic region label",
              "title": "Region"
            }
          },
          {
            "description": "Filter by co-occurrence time window in days",
            "in": "query",
            "name": "time_window_days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by co-occurrence time window in days",
              "title": "Time Window Days"
            }
          },
          {
            "description": "Filter by minimum statistical confidence",
            "in": "query",
            "name": "min_confidence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by minimum statistical confidence",
              "title": "Min Confidence"
            }
          },
          {
            "description": "Sort attribute",
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "confidence",
              "description": "Sort attribute",
              "enum": [
                "confidence",
                "support_count",
                "lift",
                "mined_at"
              ],
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "description": "Sort direction",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "description": "Sort direction",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PatternFindingResponse"
                  },
                  "title": "Response List Public Research Patterns Api V1 Research Patterns Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Public research API disabled"
          }
        },
        "summary": "List mined longitudinal research patterns",
        "tags": [
          "v1",
          "Public Research"
        ]
      }
    }
  }
}